requirements.txt
pdfminer.six
gpt_researcher
# aisuite[all]
python-dotenv>=1.0.0
matplotlib
numpy
requests>=2.31.0
sentence-transformers
scikit-learn
arxiv
pymupdf4llm
pymupdf
ollama
smolagents
docstring_parser
git+https://github.com/piromagnus/aisuite.git#egg=aisuite[all]
mermaid-py
markdown
EbookLib
latex2mathml
git+https://github.com/Moonbase59/latex2svg.git#egg=latex2svg
Pillow
scour
beautifulsoup4>=4.12.0
